refactor: batch page
